pyproject.toml
[project]
name = "image-chat-bot"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "streamlit>=1.41.1",
    "openai>=1.58.1",
    "image>=1.5.33",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/image_chat_bot"]
